finished

+++ src/main/java/com/insurance/solutions/app/services/ClientService.java
@@ -8,7 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ClientService {
@@ -16,22 +15,6 @@ public class ClientService {
     @Autowired
     private ClientRepository clientRepository;
 
-
-    public Optional<Client> findById(long id) {
-        return clientRepository.findById(id);
-    }
-
-    public void deleteAll() {
-        clientRepository.deleteAll();
-    }
-
-    public Iterable<Client> saveAll(Iterable<Client> clients) {
-        return clientRepository.saveAll(clients);
-    }
-
-    public Client save(Client client) {
-        return clientRepository.save(client);
-    }
 
     public ResponseEntity<Client> createClient(Client client) {
         if (clientRepository.existsByDniAndInsuranceCompany(client.getDni(), client.getInsuranceCompany()))
@@ -43,12 +26,10 @@ public class ClientService {
     }
 
     public ResponseEntity<?> deleteClientById(long id) {
-        return clientRepository.findById(id)
-                .map(__ -> {
-                    clientRepository.deleteById(id);
-                    return ResponseEntity.ok().build();
-                })
-                .orElseThrow(() -> new BadRequestException("El cliente con id: " + id + " no existe."));
+        if (!clientRepository.existsById(id)) throw new BadRequestException("El cliente con id: " + id + " no existe.");
+
+        clientRepository.deleteById(id);
+        return ResponseEntity.ok().build();
     }
 
     public ResponseEntity<List<Client>> findAll() {
